requirements.txt
asgiref==3.8.1
attrs==25.3.0
black==25.1.0
cfgv==3.4.0
click==8.1.8
distlib==0.3.9
Django==5.1.7
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.15.2
drf-spectacular==0.28.0
exceptiongroup==1.2.2
filelock==3.18.0
flake8==7.1.2
identify==2.6.9
inflection==0.5.1
iniconfig==2.0.0
isort==6.0.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
mccabe==0.7.0
mypy-extensions==1.0.0
nodeenv==1.9.1
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
pluggy==1.5.0
pre_commit==4.1.0
psycopg2-binary==2.9.10
pycodestyle==2.12.1
pyflakes==3.2.0
pytest==8.3.5
pytest-django==4.10.0
python-dotenv==1.0.1
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.23.1
sqlparse==0.5.3
tomli==2.2.1
typing_extensions==4.12.2
uritemplate==4.1.1
virtualenv==20.29.3
